+++ docs/changes.rst
@@ -1,7 +1,7 @@
 Weblate 4.8
--------------
+-----------
 
-Not yet released.
+Released on August 21th 2021.
 
 * Added support for Apple stringsdict format.
 * The exact search opearator is now case-sensitive on PostgreSQL.